refactor : moved batchconfig from config dir to batch dir
